refactor: code cleanup — dead code purge, file splits, TestFlight prep
Rebased cleanup phases 1-4 onto latest main (post-PR #125).

- Phase 1: zones default, remove SacHomeView, gate DevMode, fix privacy manifest
- Phase 2: dead code purge (RecordingState, SessionSummaryService, extractEntries, etc.)
- Phase 3: split LLMService.swift into AgentTypes/LLMPrompt/PPQTypes, extract pulse animation
- Phase 4: meta docs cleanup

Conflict resolution preserved main's feature additions (analytics props on
LLMConversation, notes/checkOffHabit fields, habit prompt rules) while
applying cleanup's structural changes.

Fixed DevModeView build break: removed references to deleted RecordingState
enum, recordingState property, and showFocusCard property.
