+++ Murmur/Components/EmptyStatePulseView.swift
@@ -1,0 +1,119 @@
+import SwiftUI
+
+/// Shared empty-state view with concentric pulsing circles, a mic button,
+/// and a title/subtitle message. Encapsulates the 6 `@State` animation
+/// properties and `startPulseAnimation()` that were previously duplicated
+/// across home view variants.
+struct EmptyStatePulseView: View {
+    let onMicTap: () -> Void
+    var title: String = "Say or type anything."
+    var subtitle: String = "Murmur remembers so you don't have to."
+    var showDevModeActivator: Bool = false
+
+    // Pulse animation state
+    @State private var pulseScale1: CGFloat = 1.0
+    @State private var pulseScale2: CGFloat = 1.0
+    @State private var pulseScale3: CGFloat = 1.0
+    @State private var pulseOpacity1: Double = 1.0
+    @State private var pulseOpacity2: Double = 0.7
+    @State private var pulseOpacity3: Double = 0.5
+
+    var body: some View {
+        VStack(spacing: 0) {
+            Spacer()
+
+            VStack(spacing: 40) {
+                ZStack {
+                    Circle()
+                        .stroke(Theme.Colors.accentPurple.opacity(0.05), lineWidth: 1)
+                        .frame(width: 136, height: 136)
+                        .scaleEffect(pulseScale3)
+                        .opacity(pulseOpacity3)
+
+                    Circle()
+                        .stroke(Theme.Colors.accentPurple.opacity(0.1), lineWidth: 1)
+                        .frame(width: 112, height: 112)
+                        .scaleEffect(pulseScale2)
+                        .opacity(pulseOpacity2)
+
+                    Circle()
+                        .stroke(Theme.Colors.accentPurple.opacity(0.3), lineWidth: 2)
+                        .frame(width: 88, height: 88)
+                        .scaleEffect(pulseScale1)
+                        .opacity(pulseOpacity1)
+
+                    Button(action: onMicTap) {
+                        Image(systemName: "mic")
+                            .font(.largeTitle)
+                            .foregroundStyle(Theme.Colors.accentPurple.opacity(0.5))
+                    }
+                    .buttonStyle(.plain)
+                    .accessibilityLabel("Record your first voice note")
+                }
+                .onAppear { startPulseAnimation() }
+
+                VStack(spacing: 10) {
+                    Text(title)
+                        .font(Theme.Typography.title)
+                        .tracking(-0.5)
+                        .foregroundStyle(Theme.Colors.textPrimary)
+
+                    subtitleView
+                }
+                .multilineTextAlignment(.center)
+                .padding(.horizontal, 40)
+            }
+
+            Spacer()
+            Spacer()
+        }
+    }
+
+    @ViewBuilder
+    private var subtitleView: some View {
+        let base = Text(subtitle)
+            .font(Theme.Typography.body)
+            .foregroundStyle(Theme.Colors.textSecondary)
+            .lineSpacing(2)
+
+        if showDevModeActivator {
+            #if DEBUG
+            base.devModeActivator()
+            #else
+            base
+            #endif
+        } else {
+            base
+        }
+    }
+
+    // MARK: - Pulse Animation
+
+    private func startPulseAnimation() {
+        withAnimation(
+            .easeInOut(duration: 3)
+            .repeatForever(autoreverses: true)
+        ) {
+            pulseScale1 = 1.05
+            pulseOpacity1 = 0.8
+        }
+
+        withAnimation(
+            .easeInOut(duration: 3)
+            .repeatForever(autoreverses: true)
+            .delay(0.5)
+        ) {
+            pulseScale2 = 1.05
+            pulseOpacity2 = 0.5
+        }
+
+        withAnimation(
+            .easeInOut(duration: 3)
+            .repeatForever(autoreverses: true)
+            .delay(1.0)
+        ) {
+            pulseScale3 = 1.05
+            pulseOpacity3 = 0.3
+        }
+    }
+}
