+++ Packages/MurmurCore/Package.swift
@@ -9,19 +9,9 @@ let package = Package(
     ],
     products: [
         .library(name: "MurmurCore", targets: ["MurmurCore"]),
-        .executable(name: "TranscriptionTest", targets: ["TranscriptionTest"]),
-        .executable(name: "ScenarioRunner", targets: ["ScenarioRunner"]),
     ],
     targets: [
         .target(name: "MurmurCore"),
-        .executableTarget(
-            name: "TranscriptionTest",
-            dependencies: ["MurmurCore"]
-        ),
-        .executableTarget(
-            name: "ScenarioRunner",
-            dependencies: ["MurmurCore"]
-        ),
         .testTarget(
             name: "MurmurCoreTests",
             dependencies: ["MurmurCore"]
