+++ Packages/MurmurCore/Sources/MurmurCore/LLMService.swift
@@ -57,463 +57,6 @@ public enum LLMToolChoice: Sendable {
     }
 }
 
-/// Configuration for LLM extraction/agent turns: system prompt + tool definitions.
-/// Define once, pass to any LLMService implementation.
-public struct LLMPrompt: @unchecked Sendable {
-    public let systemPrompt: String
-    public let tools: [[String: Any]]
-    public let toolChoice: LLMToolChoice
-
-    public init(systemPrompt: String, tools: [[String: Any]], toolChoice: LLMToolChoice) {
-        self.systemPrompt = systemPrompt
-        self.tools = tools
-        self.toolChoice = toolChoice
-    }
-
-    /// Agentic entry manager prompt (phase one): create/update/complete/archive tools.
-    public static let entryManager = LLMPrompt(
-        systemPrompt: """
-            You are Murmur, a personal entry manager for voice input.
-
-            You receive:
-            1) A compact list of current entries (may be empty)
-            2) New user transcript text from speech recognition (contains transcription errors)
-
-            Your job is to decide which actions to take using tools:
-            - create_entries: add genuinely new entries
-            - update_entries: modify existing entry fields (including snooze via status + snooze_until)
-            - complete_entries: mark entries done
-            - archive_entries: remove no-longer-relevant entries
-
-            Decision rules:
-            - Prefer updating/completing existing entries over creating duplicates.
-            - Use fuzzy semantic matching for references ("that one", "the dentist thing", garbled names).
-            - If user says done/finished/completed a non-habit entry, use complete_entries.
-            - For habits: when user says they did/completed their habit, use update_entries
-              with check_off_habit: true. Marks it done for the period, keeps it active.
-              Do NOT use complete_entries for habits — that archives them permanently.
-            - If user changes timing/priority/details, use update_entries.
-            - Only create when intent is genuinely new.
-            - If no current entries are provided, create_entries is usually appropriate.
-
-            Create entry quality rules:
-            - Produce concise card-style content, not long prose.
-            - summary should be 10 words or fewer.
-            - Set due_date as an ISO 8601 datetime string resolved from the current time (e.g. "2025-03-17T15:30:00"). Always set it when the user mentions any time reference.
-            - Keep snooze_until as the user's natural language phrase.
-            - For habits, set cadence to daily/weekdays/weekly/monthly when clear.
-            - Priority 1-5 (1=highest). Default 3 unless words signal urgency ("urgent", "ASAP", "critical" → 1-2) or low importance ("whenever", "someday" → 4-5).
-            - Do not include urgency words in content when priority captures urgency.
-
-            Mutation rules:
-            - Every update/complete/archive item must include a short reason.
-            - Use the provided entry id exactly as given in context.
-
-            Output rules:
-            - Use tool calls for all entry operations.
-            - After tool calls, include a brief status message (under 15 words) summarizing what you did.
-              Examples: "Added 2 reminders for tomorrow", "Marked dentist appointment complete", "Updated grocery list with 3 items"
-              This message appears as a notification to the user — be concise and specific.
-            - Call multiple tools when needed.
-            - Act decisively on clear inputs — never ask for confirmation on straightforward requests.
-            - When genuinely ambiguous (e.g., multiple entries match, or intent unclear), use confirm_actions to propose what you'd do — the user sees a preview and confirms or declines.
-
-            Memory rules:
-            - You have persistent memory across sessions via update_memory.
-            - Store: user preferences, naming patterns, recurring schedules, vocabulary corrections.
-            - Do NOT store entry data (already in context). Keep under 500 words.
-            - Replace full content each time. Only update when you learn something new.
-
-            Layout tools:
-            - get_current_layout reads the current home screen layout as JSON.
-            - update_layout applies incremental changes as an animated batch.
-            - After entry operations, call get_current_layout then update_layout to reflect changes.
-            - If the layout is empty (cold start), build it with add_section + insert_entry.
-            - Calling update_layout is optional — entries without placement appear above the layout.
-            - See ## Layout Instructions in the user message for the active layout style.
-            """,
-        tools: [
-            createEntriesToolSchema(),
-            updateEntriesToolSchema(),
-            completeEntriesToolSchema(),
-            archiveEntriesToolSchema(),
-            updateMemoryToolSchema(),
-            confirmActionsToolSchema(),
-            getCurrentLayoutToolSchema(),
-            updateLayoutToolSchema(),
-        ],
-        toolChoice: .auto
-    )
-
-    /// Backward-compatible extraction prompt for current UI flow.
-    /// Uses only create_entries so existing confirm/save UI still works.
-    public static let entryCreation = LLMPrompt(
-        systemPrompt: """
-            You are an extraction assistant for a voice-to-entries app.
-
-            The transcript comes from speech recognition and can contain errors.
-            Infer intended meaning and clean up wording.
-
-            Extract intentional items the user wants to track (todo, reminder, note,
-            idea, list, habit, question). Skip filler and small talk.
-
-            Return only create_entries tool calls. Each entry should include:
-            - content: concise card text
-            - category
-            - source_text: relevant transcript span
-            - summary: 10 words or fewer
-            Optional: priority (1-5, default 3 unless urgency/low importance is clear), \
-            due_date (ISO 8601 datetime resolved from current time, e.g. "2025-03-17T15:30:00"), cadence (for habits)
-            Always set due_date for reminders and todos when the user mentions any time reference (e.g. "in 30 minutes", "tomorrow", "next Friday").
-            """,
-        tools: [createEntriesToolSchema()],
-        toolChoice: .function(name: "create_entries")
-    )
-
-    @available(*, deprecated, message: "Use entryManager or entryCreation")
-    public static let entryExtraction = entryCreation
-
-    /// Home composition prompt: composes the entire home screen layout via compose_view tool.
-    public static let homeComposition = LLMPrompt(
-        systemPrompt: """
-            You are composing a home screen for a personal voice assistant app.
-            You receive the user's current entries. Compose 3-5 sections showing what matters RIGHT NOW.
-
-            Rules:
-            - Most entries stay hidden. Show up to 7 items total.
-            - Group by urgency and context, NOT by category.
-            - First section: what needs attention now (overdue, due today, P1/P2). Use relaxed density, hero emphasis for urgent items.
-            - Later sections: things to keep in mind, upcoming items. Use compact density.
-            - Include a brief message (under 15 words) if it adds context. Don't force one.
-            - Assign badges: "Overdue" for past-due, "Today" for due today, "Stale" for untouched 7+ days.
-            - Use hero emphasis sparingly (1-2 items max). Compact for low-priority items.
-            - If nothing is urgent, compose a calm view with a reassuring message.
-            - If no entries exist, return zero sections.
-            """,
-        tools: [composeViewToolSchema()],
-        toolChoice: .function(name: "compose_view")
-    )
-
-    /// Navigator composition prompt: category-grouped, up to 7 items, with briefing.
-    public static let navigatorComposition = LLMPrompt(
-        systemPrompt: """
-            You are composing a home screen for a personal voice assistant app.
-            You receive the user's current entries. Select up to 7 entries that deserve attention today.
-
-            Rules:
-            - One section per category (section title = category name lowercase: todo, reminder, habit, note, idea, list, question).
-            - Only include categories that have selected entries — don't create empty sections.
-            - Use relaxed density for all sections.
-            - Use standard emphasis for all entries.
-            - Badge = short human-readable reason this entry needs attention: "Overdue", "Due today", "High priority", "New", "Stale".
-            - Produce a briefing: one friendly sentence summarizing what the day looks like.
-              Example: "You have 3 things due today and a habit to maintain."
-            - No inline message items — use the briefing field instead.
-            - If nothing deserves focus, return zero sections and a calm briefing like "All clear — nothing pressing today."
-
-            Selection criteria (in priority order):
-            1. Overdue entries (due date has passed)
-            2. Due today
-            3. High priority (P1, P2)
-            4. Stale entries (created long ago, never updated)
-            5. Habits not yet done for the current period
-            """,
-        tools: [composeViewToolSchema()],
-        toolChoice: .function(name: "compose_view")
-    )
-
-    /// Layout refresh prompt: compare current layout against entries, output diffs only.
-    public static let layoutRefresh = LLMPrompt(
-        systemPrompt: """
-            You are refreshing a home screen layout for a voice assistant app.
-            You receive the current layout (JSON) and the current entries.
-            Compare them and output update_layout operations to bring the layout up to date.
-
-            Rules:
-            - Remove entries no longer active (completed, archived, deleted).
-            - Add entries that deserve attention but are missing from the layout.
-            - Update badges based on current dates: "Overdue" if past due, "Today" if due today, etc.
-            - Move entries whose urgency or context has changed.
-            - Update emphasis if priority has shifted.
-            - Preserve the overall layout structure — minimize churn.
-            - If no changes are needed, call update_layout with an empty operations array.
-            - See ## Layout Instructions for the active layout style constraints.
-
-            Always call update_layout exactly once.
-            """,
-        tools: [updateLayoutToolSchema()],
-        toolChoice: .function(name: "update_layout")
-    )
-}
-
-/// Agent-facing status for context snapshots and update fields.
-public enum AgentEntryStatus: String, Codable, Sendable, CaseIterable {
-    case active
-    case completed
-    case archived
-    case snoozed
-}
-
-/// Compact entry snapshot passed into the LLM context.
-public struct AgentContextEntry: Sendable, Codable, Identifiable {
-    public let id: String
-    public let summary: String
-    public let category: EntryCategory
-    public let priority: Int?
-    public let dueDateDescription: String?
-    public let cadence: HabitCadence?
-    public let status: AgentEntryStatus
-    public let createdAt: Date
-    public let currentStreak: Int?
-    public let notes: String?
-
-    public init(
-        id: String,
-        summary: String,
-        category: EntryCategory,
-        priority: Int? = nil,
-        dueDateDescription: String? = nil,
-        cadence: HabitCadence? = nil,
-        status: AgentEntryStatus = .active,
-        createdAt: Date = Date(),
-        currentStreak: Int? = nil,
-        notes: String? = nil
-    ) {
-        self.id = id
-        self.summary = summary
-        self.category = category
-        self.priority = priority
-        self.dueDateDescription = dueDateDescription
-        self.cadence = cadence
-        self.status = status
-        self.createdAt = createdAt
-        self.currentStreak = currentStreak
-        self.notes = notes
-    }
-}
-
-public struct CreateAction: Sendable {
-    public let content: String
-    public let category: EntryCategory
-    public let sourceText: String
-    public let summary: String
-    public let priority: Int?
-    public let dueDateDescription: String?
-    public let cadence: HabitCadence?
-    public let notes: String?
-
-    public init(
-        content: String,
-        category: EntryCategory,
-        sourceText: String,
-        summary: String,
-        priority: Int? = nil,
-        dueDateDescription: String? = nil,
-        cadence: HabitCadence? = nil,
-        notes: String? = nil
-    ) {
-        self.content = content
-        self.category = category
-        self.sourceText = sourceText
-        self.summary = summary
-        self.priority = priority
-        self.dueDateDescription = dueDateDescription
-        self.cadence = cadence
-        self.notes = notes
-    }
-}
-
-public struct UpdateFields: Sendable {
-    public let content: String?
-    public let summary: String?
-    public let category: EntryCategory?
-    public let priority: Int?
-    public let dueDateDescription: String?
-    public let cadence: HabitCadence?
-    public let status: AgentEntryStatus?
-    public let snoozeUntilDescription: String?
-    public let checkOffHabit: Bool?
-    public let notes: String?
-
-    public init(
-        content: String? = nil,
-        summary: String? = nil,
-        category: EntryCategory? = nil,
-        priority: Int? = nil,
-        dueDateDescription: String? = nil,
-        cadence: HabitCadence? = nil,
-        status: AgentEntryStatus? = nil,
-        snoozeUntilDescription: String? = nil,
-        checkOffHabit: Bool? = nil,
-        notes: String? = nil
-    ) {
-        self.content = content
-        self.summary = summary
-        self.category = category
-        self.priority = priority
-        self.dueDateDescription = dueDateDescription
-        self.cadence = cadence
-        self.status = status
-        self.snoozeUntilDescription = snoozeUntilDescription
-        self.checkOffHabit = checkOffHabit
-        self.notes = notes
-    }
-}
-
-public struct UpdateAction: Sendable {
-    public let id: String
-    public let fields: UpdateFields
-    public let reason: String
-
-    public init(id: String, fields: UpdateFields, reason: String) {
-        self.id = id
-        self.fields = fields
-        self.reason = reason
-    }
-}
-
-public struct CompleteAction: Sendable {
-    public let id: String
-    public let reason: String
-
-    public init(id: String, reason: String) {
-        self.id = id
-        self.reason = reason
-    }
-}
-
-public struct ArchiveAction: Sendable {
-    public let id: String
-    public let reason: String
-
-    public init(id: String, reason: String) {
-        self.id = id
-        self.reason = reason
-    }
-}
-
-public struct UpdateMemoryAction: Sendable {
-    public let content: String
-
-    public init(content: String) {
-        self.content = content
-    }
-}
-
-/// Proposed actions awaiting user confirmation.
-public struct ConfirmationRequest: Sendable {
-    public let message: String
-    public let proposedActions: [AgentAction]
-
-    public init(message: String, proposedActions: [AgentAction]) {
-        self.message = message
-        self.proposedActions = proposedActions
-    }
-}
-
-/// Typed actions produced by the agent.
-public enum AgentAction: Sendable {
-    case create(CreateAction)
-    case update(UpdateAction)
-    case complete(CompleteAction)
-    case archive(ArchiveAction)
-    case updateMemory(UpdateMemoryAction)
-    case confirm(ConfirmationRequest)
-    case layoutRead
-    case layoutUpdate([LayoutOperation])
-}
-
-public extension AgentAction {
-    var createdEntry: ExtractedEntry? {
-        guard case .create(let action) = self else {
-            return nil
-        }
-
-        return ExtractedEntry(
-            content: action.content,
-            category: action.category,
-            sourceText: action.sourceText,
-            summary: action.summary,
-            priority: action.priority,
-            dueDateDescription: action.dueDateDescription,
-            cadence: action.cadence,
-            notes: action.notes
-        )
-    }
-
-    var isConfirmation: Bool {
-        if case .confirm = self { return true }
-        return false
-    }
-
-    /// The entry ID targeted by a mutation action (update/complete/archive), or nil for creates/other.
-    var mutationEntryID: String? {
-        switch self {
-        case .update(let a): return a.id
-        case .complete(let a): return a.id
-        case .archive(let a): return a.id
-        default: return nil
-        }
-    }
-}
-
-/// A tool call that failed to decode.
-public struct ParseFailure: Sendable {
-    public let toolName: String
-    public let rawArguments: String
-    public let errorDescription: String
-    public let toolCallID: String?
-
-    public init(toolName: String, rawArguments: String, errorDescription: String, toolCallID: String? = nil) {
-        self.toolName = toolName
-        self.rawArguments = rawArguments
-        self.errorDescription = errorDescription
-        self.toolCallID = toolCallID
-    }
-}
-
-/// Maps a single LLM tool_call_id to the contiguous slice of actions it produced.
-public struct ToolCallGroup: Sendable {
-    public let toolCallID: String
-    public let toolName: String
-    public let actionRange: Range<Int>
-
-    public init(toolCallID: String, toolName: String, actionRange: Range<Int>) {
-        self.toolCallID = toolCallID
-        self.toolName = toolName
-        self.actionRange = actionRange
-    }
-}
-
-/// The agent response for a single turn.
-public struct AgentResponse: Sendable {
-    public let actions: [AgentAction]
-    public let summary: String
-    public let usage: TokenUsage
-    public let parseFailures: [ParseFailure]
-    public let toolCallGroups: [ToolCallGroup]
-
-    /// Text content from the agent when responding without tool calls (clarifications, summaries).
-    /// Only populated when the model returns text content — nil when actions-only.
-    public let textResponse: String?
-
-    public init(
-        actions: [AgentAction],
-        summary: String,
-        usage: TokenUsage,
-        parseFailures: [ParseFailure] = [],
-        toolCallGroups: [ToolCallGroup] = [],
-        textResponse: String? = nil
-    ) {
-        self.actions = actions
-        self.summary = summary
-        self.usage = usage
-        self.parseFailures = parseFailures
-        self.toolCallGroups = toolCallGroups
-        self.textResponse = textResponse
-    }
-}
-
 /// Protocol for the entry-management agent.
 public protocol MurmurAgent: Sendable {
     func process(
@@ -530,444 +73,4 @@ public protocol StreamingMurmurAgent: MurmurAgent {
         existingEntries: [AgentContextEntry],
         conversation: LLMConversation
     ) -> AsyncThrowingStream<AgentStreamEvent, Error>
-}
-
-/// LLM-backed implementation contract.
-/// Keeps extractEntries for current UI compatibility.
-public protocol LLMService: MurmurAgent {
-    /// Extract entries from a transcript with auto-categorization.
-    /// This is a compatibility method for current extraction-first UI flows.
-    func extractEntries(from transcript: String, conversation: LLMConversation) async throws -> LLMResult
-}
-
-public extension LLMService {
-    func extractEntries(from transcript: String, conversation: LLMConversation) async throws -> LLMResult {
-        let response = try await process(
-            transcript: transcript,
-            existingEntries: [],
-            conversation: conversation
-        )
-
-        return LLMResult(
-            entries: response.actions.compactMap(\.createdEntry),
-            usage: response.usage
-        )
-    }
-}
-
-public struct LLMResult: Sendable {
-    public let entries: [ExtractedEntry]
-    public let usage: TokenUsage
-
-    public init(entries: [ExtractedEntry], usage: TokenUsage) {
-        self.entries = entries
-        self.usage = usage
-    }
-}
-
-/// An entry extracted by the LLM — pure value type, no persistence dependency.
-public struct ExtractedEntry: Sendable, Codable, Identifiable {
-    /// Stable identity for SwiftUI (not decoded from JSON — always freshly generated)
-    public let id: UUID
-
-    /// The processed/cleaned content
-    public let content: String
-
-    /// The category assigned by the LLM
-    public let category: EntryCategory
-
-    /// The specific part of the transcript this was extracted from
-    public let sourceText: String
-
-    /// One-liner summary for cards/lists
-    public let summary: String
-
-    /// Priority 1-5 scale (1 = highest), nil if not applicable
-    public let priority: Int?
-
-    /// Raw time phrase from transcript (e.g. "next Thursday"), nil if none
-    public let dueDateDescription: String?
-
-    /// How often this habit repeats, nil for non-habits
-    public let cadence: HabitCadence?
-
-    /// Supplementary notes for the entry
-    public let notes: String?
-
-    enum CodingKeys: String, CodingKey {
-        case content, category, sourceText, summary, priority, dueDateDescription, cadence, notes
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.id = UUID()
-        self.content = try container.decode(String.self, forKey: .content)
-        self.category = try container.decode(EntryCategory.self, forKey: .category)
-        self.sourceText = try container.decode(String.self, forKey: .sourceText)
-        self.summary = try container.decodeIfPresent(String.self, forKey: .summary) ?? ""
-        self.priority = try container.decodeIfPresent(Int.self, forKey: .priority)
-        self.dueDateDescription = try container.decodeIfPresent(String.self, forKey: .dueDateDescription)
-        self.cadence = try container.decodeIfPresent(HabitCadence.self, forKey: .cadence)
-        self.notes = try container.decodeIfPresent(String.self, forKey: .notes)
-    }
-
-    public init(
-        content: String,
-        category: EntryCategory,
-        sourceText: String,
-        summary: String = "",
-        priority: Int? = nil,
-        dueDateDescription: String? = nil,
-        cadence: HabitCadence? = nil,
-        notes: String? = nil
-    ) {
-        self.id = UUID()
-        self.content = content
-        self.category = category
-        self.sourceText = sourceText
-        self.summary = summary
-        self.priority = priority
-        self.dueDateDescription = dueDateDescription
-        self.cadence = cadence
-        self.notes = notes
-    }
-}
-
-private extension LLMPrompt {
-    static func createEntriesToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "create_entries",
-                "description": "Create new entries from user intent",
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "entries": [
-                            "type": "array",
-                            "items": [
-                                "type": "object",
-                                "properties": [
-                                    "content": [
-                                        "type": "string",
-                                        "description": "Cleaned, concise entry content",
-                                    ],
-                                    "category": [
-                                        "type": "string",
-                                        "enum": ["todo", "note", "reminder", "idea", "list", "habit", "question"],
-                                    ],
-                                    "source_text": [
-                                        "type": "string",
-                                        "description": "Relevant source span from transcript",
-                                    ],
-                                    "summary": [
-                                        "type": "string",
-                                        "description": "Card title, 10 words or fewer",
-                                    ],
-                                    "priority": ["type": "integer"],
-                                    "due_date": ["type": "string"],
-                                    "cadence": [
-                                        "type": "string",
-                                        "enum": ["daily", "weekdays", "weekly", "monthly"],
-                                    ],
-                                    "notes": [
-                                        "type": "string",
-                                        "description": "Supplementary notes for the entry",
-                                    ] as [String: Any],
-                                ] as [String: Any],
-                                "required": ["content", "category", "source_text", "summary"],
-                            ] as [String: Any],
-                        ] as [String: Any],
-                    ],
-                    "required": ["entries"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func updateEntriesToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "update_entries",
-                "description": "Update one or more existing entries",
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "updates": [
-                            "type": "array",
-                            "items": [
-                                "type": "object",
-                                "properties": [
-                                    "id": ["type": "string"],
-                                    "fields": [
-                                        "type": "object",
-                                        "properties": [
-                                            "content": ["type": "string"],
-                                            "summary": ["type": "string"],
-                                            "category": [
-                                                "type": "string",
-                                                "enum": ["todo", "note", "reminder", "idea", "list", "habit", "question"],
-                                            ],
-                                            "priority": ["type": "integer"],
-                                            "due_date": ["type": "string"],
-                                            "cadence": [
-                                                "type": "string",
-                                                "enum": ["daily", "weekdays", "weekly", "monthly"],
-                                            ],
-                                            "status": [
-                                                "type": "string",
-                                                "enum": ["active", "snoozed", "completed", "archived"],
-                                            ],
-                                            "snooze_until": ["type": "string"],
-                                            "check_off_habit": [
-                                                "type": "boolean",
-                                                "description": "Mark a habit as done for its current period (daily/weekly/monthly). Do NOT use complete_entries for habits.",
-                                            ] as [String: Any],
-                                            "notes": [
-                                                "type": "string",
-                                                "description": "Supplementary notes for the entry",
-                                            ] as [String: Any],
-                                        ] as [String: Any],
-                                    ],
-                                    "reason": [
-                                        "type": "string",
-                                        "description": "Why this update is being applied",
-                                    ],
-                                ] as [String: Any],
-                                "required": ["id", "fields", "reason"],
-                            ] as [String: Any],
-                        ] as [String: Any],
-                    ],
-                    "required": ["updates"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func completeEntriesToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "complete_entries",
-                "description": "Mark one or more existing entries as completed",
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "entries": [
-                            "type": "array",
-                            "items": [
-                                "type": "object",
-                                "properties": [
-                                    "id": ["type": "string"],
-                                    "reason": ["type": "string"],
-                                ] as [String: Any],
-                                "required": ["id", "reason"],
-                            ] as [String: Any],
-                        ] as [String: Any],
-                    ],
-                    "required": ["entries"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func archiveEntriesToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "archive_entries",
-                "description": "Archive one or more existing entries",
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "entries": [
-                            "type": "array",
-                            "items": [
-                                "type": "object",
-                                "properties": [
-                                    "id": ["type": "string"],
-                                    "reason": ["type": "string"],
-                                ] as [String: Any],
-                                "required": ["id", "reason"],
-                            ] as [String: Any],
-                        ] as [String: Any],
-                    ],
-                    "required": ["entries"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func getCurrentLayoutToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "get_current_layout",
-                "description": "Read the current home screen layout. Returns sections with their entries, emphasis levels, and badges. Call this before update_layout to understand what's on screen.",
-                "parameters": [
-                    "type": "object",
-                    "properties": [:] as [String: Any],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func updateLayoutToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "update_layout",
-                "description": """
-                    Apply incremental changes to the home screen layout. Operations are applied in order \
-                    as a single animated transaction. Use after create_entries/complete_entries/update_entries \
-                    to place or remove entries on screen. For a fresh layout (cold start), use a batch of \
-                    add_section + insert_entry operations.
-                    """,
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "operations": [
-                            "type": "array",
-                            "items": [
-                                "type": "object",
-                                "properties": [
-                                    "op": [
-                                        "type": "string",
-                                        "enum": ["add_section", "remove_section", "update_section",
-                                                 "insert_entry", "remove_entry", "move_entry", "update_entry"],
-                                    ],
-                                    "title": ["type": "string", "description": "Section title (for section ops)"],
-                                    "density": ["type": "string", "enum": ["compact", "relaxed"]],
-                                    "position": ["type": "integer", "description": "0-indexed position (optional, omit to append)"],
-                                    "new_title": ["type": "string", "description": "New title for update_section"],
-                                    "entry_id": ["type": "string", "description": "Entry short ID (for entry ops)"],
-                                    "section": ["type": "string", "description": "Target section title (for insert_entry)"],
-                                    "to_section": ["type": "string", "description": "Destination section (for move_entry)"],
-                                    "to_position": ["type": "integer", "description": "Destination position (for move_entry)"],
-                                    "emphasis": ["type": "string", "enum": ["hero", "standard", "compact"]],
-                                    "badge": ["type": "string", "description": "Badge text: Overdue, Today, New, Stale, etc."],
-                                ] as [String: Any],
-                                "required": ["op"],
-                            ] as [String: Any],
-                        ] as [String: Any],
-                    ],
-                    "required": ["operations"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func composeViewToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "compose_view",
-                "description": "Compose the home view. Surface what matters right now. Most entries stay hidden. Group by urgency/context, not category. 3-5 sections max, up to 7 total items.",
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "sections": [
-                            "type": "array",
-                            "items": [
-                                "type": "object",
-                                "properties": [
-                                    "title": ["type": "string"],
-                                    "density": [
-                                        "type": "string",
-                                        "enum": ["compact", "relaxed"],
-                                    ],
-                                    "items": [
-                                        "type": "array",
-                                        "items": [
-                                            "type": "object",
-                                            "properties": [
-                                                "type": [
-                                                    "type": "string",
-                                                    "enum": ["entry", "message"],
-                                                ],
-                                                "id": ["type": "string"],
-                                                "emphasis": [
-                                                    "type": "string",
-                                                    "enum": ["hero", "standard", "compact"],
-                                                ],
-                                                "badge": ["type": "string"],
-                                                "text": ["type": "string"],
-                                            ] as [String: Any],
-                                            "required": ["type"],
-                                        ] as [String: Any],
-                                    ] as [String: Any],
-                                ] as [String: Any],
-                                "required": ["items"],
-                            ] as [String: Any],
-                        ] as [String: Any],
-                        "briefing": [
-                            "type": "string",
-                            "description": "One friendly sentence summarizing the day",
-                        ] as [String: Any],
-                    ],
-                    "required": ["sections"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func updateMemoryToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "update_memory",
-                "description": "Replace your persistent memory with updated content. Called when you learn something new about the user.",
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "content": [
-                            "type": "string",
-                            "description": "Full replacement text for your memory. Keep under 500 words.",
-                        ],
-                    ],
-                    "required": ["content"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
-
-    static func confirmActionsToolSchema() -> [String: Any] {
-        [
-            "type": "function",
-            "function": [
-                "name": "confirm_actions",
-                "description": "Propose actions for user confirmation when intent is ambiguous. The user sees a preview and confirms or declines.",
-                "parameters": [
-                    "type": "object",
-                    "properties": [
-                        "message": [
-                            "type": "string",
-                            "description": "Brief explanation of the ambiguity, under 20 words",
-                        ],
-                        "actions": [
-                            "type": "array",
-                            "description": "Proposed actions to preview",
-                            "items": [
-                                "type": "object",
-                                "properties": [
-                                    "tool": [
-                                        "type": "string",
-                                        "enum": ["create_entries", "update_entries", "complete_entries", "archive_entries"],
-                                        "description": "Which tool to call if confirmed",
-                                    ],
-                                    "arguments": [
-                                        "type": "object",
-                                        "description": "Arguments in the same format as the respective tool",
-                                    ],
-                                ] as [String: Any],
-                                "required": ["tool", "arguments"],
-                            ] as [String: Any],
-                        ] as [String: Any],
-                    ] as [String: Any],
-                    "required": ["message", "actions"],
-                ] as [String: Any],
-            ] as [String: Any],
-        ]
-    }
 }
