+++ Packages/MurmurCore/Tests/MurmurCoreTests/PPQLLMServiceTests.swift
@@ -4,168 +4,7 @@ import Testing
 
 @Suite("PPQLLMService", .serialized)
 struct PPQLLMServiceTests {
-    // MARK: - Response Parsing via MockURLProtocol
-
-    @Test("Parses tool call response into ExtractedEntries")
-    func parsesToolCallResponse() async throws {
-        let args = """
-        {"entries":[\
-        {"content":"Buy groceries","category":"todo",\
-        "source_text":"I need to buy groceries",\
-        "summary":"Get groceries","priority":3},\
-        {"content":"Meeting moved to Thursday","category":"reminder",\
-        "source_text":"my meeting got moved to Thursday",\
-        "summary":"Meeting rescheduled","due_date":"Thursday"}]}
-        """.trimmingCharacters(in: .whitespacesAndNewlines)
-        let escapedArgs = args.replacingOccurrences(of: "\"", with: "\\\"")
-        let responseJSON = """
-            {
-                "id": "chatcmpl-123",
-                "object": "chat.completion",
-                "choices": [{
-                    "index": 0,
-                    "message": {
-                        "role": "assistant",
-                        "content": null,
-                        "tool_calls": [{
-                            "id": "call_abc",
-                            "type": "function",
-                            "function": {
-                                "name": "create_entries",
-                                "arguments": "\(escapedArgs)"
-                            }
-                        }]
-                    },
-                    "finish_reason": "stop"
-                }]
-            }
-            """
-
-        let (service, _) = makeService(responseBody: responseJSON, statusCode: 200)
-        let result = try await service.extractEntries(
-            from: "I need to buy groceries and my meeting got moved to Thursday",
-            conversation: LLMConversation()
-        )
-        let entries = result.entries
-
-        #expect(entries.count == 2)
-
-        #expect(entries[0].content == "Buy groceries")
-        #expect(entries[0].category == .todo)
-        #expect(entries[0].sourceText == "I need to buy groceries")
-        #expect(entries[0].summary == "Get groceries")
-        #expect(entries[0].priority == 3)
-        #expect(entries[0].dueDateDescription == nil)
-
-        #expect(entries[1].content == "Meeting moved to Thursday")
-        #expect(entries[1].category == .reminder)
-        #expect(entries[1].sourceText == "my meeting got moved to Thursday")
-        #expect(entries[1].summary == "Meeting rescheduled")
-        #expect(entries[1].priority == nil)
-        #expect(entries[1].dueDateDescription == "Thursday")
-        #expect(result.usage == .zero)
-    }
-
-    @Test("Handles all entry categories")
-    func allCategories() async throws {
-        let args = """
-        {"entries":[\
-        {"content":"Do laundry","category":"todo","source_text":"do laundry"},\
-        {"content":"Wifi password is abc123","category":"note","source_text":"wifi password"},\
-        {"content":"Dentist on Friday","category":"reminder","source_text":"dentist friday"},\
-        {"content":"App for tracking plants","category":"idea","source_text":"app for plants"},\
-        {"content":"Eggs, bread, butter","category":"list","source_text":"eggs bread butter"},\
-        {"content":"Meditate every morning","category":"habit","source_text":"meditate morning"},\
-        {"content":"Capital of Portugal?","category":"question","source_text":"capital portugal"}]}
-        """.trimmingCharacters(in: .whitespacesAndNewlines)
-        let escapedArgs = args.replacingOccurrences(of: "\"", with: "\\\"")
-        let responseJSON = """
-            {
-                "choices": [{
-                    "message": {
-                        "tool_calls": [{
-                            "function": {
-                                "name": "create_entries",
-                                "arguments": "\(escapedArgs)"
-                            }
-                        }]
-                    }
-                }]
-            }
-            """
-
-        let (service, _) = makeService(responseBody: responseJSON, statusCode: 200)
-        let result = try await service.extractEntries(from: "test", conversation: LLMConversation())
-        let entries = result.entries
-
-        #expect(entries.count == 7)
-        #expect(entries[0].category == .todo)
-        #expect(entries[1].category == .note)
-        #expect(entries[2].category == .reminder)
-        #expect(entries[3].category == .idea)
-        #expect(entries[4].category == .list)
-        #expect(entries[5].category == .habit)
-        #expect(entries[6].category == .question)
-    }
-
-    @Test("Missing optional fields default gracefully")
-    func missingOptionalFields() async throws {
-        let responseJSON = """
-            {
-                "choices": [{
-                    "message": {
-                        "tool_calls": [{
-                            "function": {
-                                "name": "create_entries",
-                                "arguments": "{\\"entries\\":[{\\"content\\":\\"Buy milk\\",\\"category\\":\\"todo\\",\\"source_text\\":\\"buy milk\\"}]}"
-                            }
-                        }]
-                    }
-                }]
-            }
-            """
-
-        let (service, _) = makeService(responseBody: responseJSON, statusCode: 200)
-        let result = try await service.extractEntries(from: "buy milk", conversation: LLMConversation())
-        let entries = result.entries
-
-        #expect(entries.count == 1)
-        #expect(entries[0].content == "Buy milk")
-        #expect(entries[0].category == .todo)
-        #expect(entries[0].summary == "")
-        #expect(entries[0].priority == nil)
-        #expect(entries[0].dueDateDescription == nil)
-    }
-
-    @Test("Throws on HTTP error")
-    func httpError() async throws {
-        let (service, _) = makeService(
-            responseBody: "{\"error\": \"unauthorized\"}",
-            statusCode: 401
-        )
-
-        await #expect(throws: PPQError.self) {
-            try await service.extractEntries(from: "test", conversation: LLMConversation())
-        }
-    }
-
-    @Test("Returns empty entries when LLM responds with text only (no tool calls)")
-    func noToolCalls() async throws {
-        let responseJSON = """
-            {
-                "choices": [{
-                    "message": {
-                        "role": "assistant",
-                        "content": "I extracted some entries."
-                    }
-                }]
-            }
-            """
-
-        let (service, _) = makeService(responseBody: responseJSON, statusCode: 200)
-        let result = try await service.extractEntries(from: "test", conversation: LLMConversation())
-        #expect(result.entries.isEmpty)
-    }
+    // MARK: - Agent Process Tests
 
     @Test("Agent process returns text summary when no tool calls")
     func agentTextOnlyResponse() async throws {
@@ -188,51 +27,6 @@ struct PPQLLMServiceTests {
         )
         #expect(response.actions.isEmpty)
         #expect(response.summary == "I didn't find any actionable items in that.")
-    }
-
-    @Test("Sends correct request structure")
-    func requestStructure() async throws {
-        let responseJSON = """
-            {
-                "choices": [{
-                    "message": {
-                        "tool_calls": [{
-                            "function": {
-                                "name": "create_entries",
-                                "arguments": "{\\"entries\\":[]}"
-                            }
-                        }]
-                    }
-                }]
-            }
-            """
-
-        let (service, delegate) = makeService(responseBody: responseJSON, statusCode: 200)
-        _ = try await service.extractEntries(from: "Buy milk", conversation: LLMConversation())
-
-        // Verify request was sent correctly
-        let request = try #require(delegate.lastRequest)
-        #expect(request.url?.absoluteString == "https://api.ppq.ai/chat/completions")
-        #expect(request.httpMethod == "POST")
-        #expect(request.value(forHTTPHeaderField: "Authorization") == "Bearer test-key")
-        #expect(request.value(forHTTPHeaderField: "Content-Type") == "application/json")
-
-        // Verify request body contains expected fields
-        let body = try #require(delegate.lastRequestBody)
-        let json = try #require(JSONSerialization.jsonObject(with: body) as? [String: Any])
-        #expect(json["model"] as? String == "anthropic/claude-sonnet-4.6")
-
-        let messages = try #require(json["messages"] as? [[String: Any]])
-        #expect(messages.count == 2)
-        #expect(messages[0]["role"] as? String == "system")
-        #expect(messages[1]["role"] as? String == "user")
-        #expect(messages[1]["content"] as? String == "Buy milk")
-
-        // Verify tools are present
-        let tools = try #require(json["tools"] as? [[String: Any]])
-        #expect(tools.count == 1)
-        let function = (tools[0]["function"] as? [String: Any])
-        #expect(function?["name"] as? String == "create_entries")
     }
 
     @Test("Parses agent tools into typed update/complete/archive actions")
